PCWEB-9146 Env 라벨을 분리
